added support for no feedproducts & algorithm renaming V2 to V1
